MNT: cleanup - less ugly or more?
